+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # React Native Module Changelog
 
+## Version 14.6.0 - December 5, 2022
+Minor release that updates the iOS SDK to 16.10.6 and add support
+for isChannelCaptureEnabled and suppressAllowListError in the config for takeOff.
+
+### Changes
+- Updated iOS SDK to 16.10.6
+- Add support for **isChannelCaptureEnabled** and **suppressAllowListError** in the config for takeOff.
+
+
 ## Version 14.5.1 - November 17, 2022
 Patch release that updates the iOS SDK to 16.10.3.